+++ items.js
@@ -1,0 +1,9 @@
+const addShovel = () => {
+  $("#broken_shovel").removeClass("d-none");
+};
+
+const addFamiliarCloth = () => {
+  $("#familiar_cloth").removeClass("d-none");
+};
+
+export { addShovel, addFamiliarCloth };
